// src/api/generated/permissions.ts
export const PERMISSIONS = [
  "billing:charge",
  "billing:view",
  "customer:create",
  "customer:delete",
  "customer:edit",
  "customer:view",
  "delivery:manage",
  "delivery:pack",
  "delivery:view",
  "dispatch:send",
  "dispatch:view",
  "product:manage",
  "product:view",
  "subscription:manage",
  "subscription:view",
  "user:create",
  "user:delete",
  "user:edit",
  "user:view"
] as const;

export type Permission = (typeof PERMISSIONS)[number];